requirements.txt
django
# django-web3-auth
django-social-share
cryptography 
django-cleanup
khayyam
django-resized
django-tinymce
djangorestframework
mysqlclient
openpyxl
pusher
pylint
PyMySQL
qrcode
requests
pylint-django
django-allauth
python-dateutil
python-resize-image
Pillow
django-filter 
markdown